pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mytech_rest_auth"
version = "1.1.11"
description = "App de autenticação e gestão de entidades para Django REST Framework"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
  { name = "Dias Metano Salvador Chavana", email = "metanochava@gmail.com" }
]

dependencies = [
  "Django==5.2.10",
  "djangorestframework==3.16.1",
  "djangorestframework_simplejwt==5.5.1",
  "rest-framework-simplejwt==0.0.2",
  "drf-writable-nested==0.7.2",
  "Pillow",
  "pyotp",
  "qrcode[pil]==7.4",
  "python-barcode==0.15.1",

]

[project.urls]
Homepage = "https://github.com/metanochava/ds"
Repository = "https://github.com/metanochava/ds"

[tool.setuptools.packages.find]
where = ["src"]


